+++ src/common/profile.ui.ts
@@ -4897,7 +4897,7 @@ ipcRenderer.on('profile-item-removed', (event, type, profile, idx) => {
     $('#btn-refresh').addClass('btn-warning');
 });
 
-ipcRenderer.on('profile-updated', (event, profile, noChanges) => {
+ipcRenderer.on('profile-updated', (event, profile, noChanges, type) => {
     //filesChanged = true;
     //$('#btn-refresh').addClass('btn-warning');
 });